refactor(graduate): 각 도메인 사용처로 api 메서드 이동 및 분리
