+++ apps/graduate/src/admin/pages/filePreview/api/index.ts
@@ -1,0 +1,2 @@
+export * from './fetchCertificateFile';
+export * from './fetchThesisFile';
