--- apps/graduate/src/admin/pages/filePreview/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useFile } from './useFile';
