+++ apps/graduate/src/admin/pages/filePreview/model/file.ts
