+++ apps/graduate/src/admin/pages/filePreview/model/index.ts
@@ -1,0 +1,1 @@
+export * from './useFile';
