+++ apps/graduate/src/admin/pages/filePreview/utils/transformCertificateFileResponse.ts
@@ -1,4 +1,4 @@
-import { FileItem } from '../model/filePreview';
+import { type FileItem } from '../model/file';
 
 import { CertificateFileResponse } from '~/admin/shared/types/file';
