+++ apps/graduate/src/admin/shared/api/index.ts
@@ -1,3 +1,2 @@
 export * from './schedule';
 export * from './notice';
-export * from './file';
